+++ execution/executor/send.go
@@ -21,10 +21,10 @@ func (e *SendExecutor) Execute(trx *tx.Tx, sb sandbox.Sandbox) error {
 	pld := trx.Payload().(*payload.SendPayload)
 
 	if !e.strict && trx.IsSubsidyTx() {
-		// In not-strict mode we accepts all subsidy transactions for the current height
+		// In not-strict mode all subsidy transactions for the current height are valid.
 		// There might be more than one valid subsidy transaction per height,
-		// because There might be more than one proposal per height.
-		if trx.Sequence() != sb.CurrentHeight() {
+		// because there might be more than one proposal per height.
+		if uint32(trx.Sequence()) != sb.CurrentHeight() {
 			return errors.Errorf(errors.ErrInvalidSequence,
 				"subsidy transaction is not for current height, expected :%d, got: %d",
 				sb.CurrentHeight(), trx.Sequence())
